migrate: move files to `packages/core` - auto updates imports - creates minimal package.json and tsconfig - does not compile
