+++ packages/core/package.json
@@ -1,0 +1,6 @@
+{
+  "name": "@mergenb/core",
+  "version": "0.0.0",
+  "private": true,
+  "type": "commonjs"
+}
