+++ packages/core/src/notebookUtils.ts
