+++ packages/core/src/types/katexAutoRender.d.ts
