+++ src/tests/playwright/syntaxHighlighting.spec.ts
@@ -5,7 +5,7 @@
  */
 
 import { test, expect } from './fixtures';
-import * as logger from '../../logger';
+import * as logger from '../../../packages/core/src/logger';
 import {
     readSettingsFileSnapshot,
     restoreSettingsFileSnapshot,
