+++ src/tests/playwright/takeAllButtons.spec.ts
@@ -8,7 +8,7 @@
 
 import { execFileSync } from 'child_process';
 import { test, expect } from './fixtures';
-import * as logger from '../../logger';
+import * as logger from '../../../packages/core/src/logger';
 import {
     getCellSource,
     validateNotebookStructure,
