+++ src/tests/playwright/undoRedoActions.spec.ts
@@ -8,7 +8,7 @@
 
 import type { Locator } from 'playwright';
 import { test, expect } from './fixtures';
-import * as logger from '../../logger';
+import * as logger from '../../../packages/core/src/logger';
 import {
     clickHistoryUndo,
     clickHistoryRedo,
