+++ src/tests/playwright/unmatchGuardNoReorder.spec.ts
@@ -8,7 +8,7 @@
 import type { Page } from 'playwright';
 import { test, expect } from './fixtures';
 import { waitForResolvedCount } from '../integrationUtils';
-import * as logger from '../../logger';
+import * as logger from '../../../packages/core/src/logger';
 import {
     readSettingsFileSnapshot,
     restoreSettingsFileSnapshot,
