+++ src/tests/unmergedStatusMatrix.test.ts
@@ -5,7 +5,7 @@ import * as gitIntegration from '../gitIntegration';
 import { NotebookConflictResolver } from '../resolver';
 import { readTestConfig } from './testHarness';
 import { git, gitAllowFailure, hashBlob } from './gitTestUtils';
-import * as logger from '../logger';
+import * as logger from '../../packages/core/src/logger';
 
 type GitStage = '1' | '2' | '3';
 type StatusSpec = {
