+++ src/web/client/index.tsx
@@ -7,7 +7,7 @@ import React from 'react';
 import { createRoot } from 'react-dom/client';
 import { App } from './App';
 import { injectStyles } from './styles';
-import * as logger from '../../logger';
+import * as logger from '../../../packages/core/src/logger';
 
 // Use server-provided theme (via data-theme attribute on #root) so loading and app boot with the same palette.
 const rootEl = document.getElementById('root');
