+++ src/web/webServer.ts
@@ -18,7 +18,7 @@ import * as path from 'path';
 import * as fs from 'fs';
 import { randomBytes, randomUUID } from 'crypto';
 import WebSocket, { WebSocketServer } from 'ws';
-import * as logger from '../logger';
+import * as logger from '../../packages/core/src/logger';
 import type { WebConflictData } from './webTypes';
 
 // VSCode is optional - only needed for openExternal
